requirements.txt
openai
beautifulsoup4
requests
schedule
google-generativeai
python-dotenv
undetected-chromedriver
setuptools
